Refactored commands and added unused package
