+++ src/main/java/javacake/Main.java
@@ -7,17 +7,14 @@ import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.image.Image;
-import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.GridPane;
 import javafx.stage.Stage;
-import javafx.stage.StageStyle;
 
 /**
  * A GUI for Duke using FXML.
  */
 public class Main extends Application {
 
-    private static String savedDataPath = "data/";
     private JavaCake javaCake = new JavaCake();
