+++ src/main/java/javacake/storage/TaskList.java
@@ -1,15 +1,10 @@
 package javacake.storage;
 
-import com.joestelmach.natty.DateGroup;
 import javacake.exceptions.CakeException;
 import javacake.tasks.Deadline;
-import javacake.tasks.RecurringTask;
 import javacake.tasks.Task;
-import javacake.tasks.ToDo;
 
 import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
 
 public class TaskList {
     private ArrayList<Task> data;
@@ -79,23 +74,6 @@ public class TaskList {
     }
 
     /**
-     * Creates a new 'toBeDone' task, before adding it to current list,
-     * then returning the output by Duke.
-     * @param data ArrayList of Tasks that's currently being stored
-     * @param input Command input by user
-     * @param state The type of output needed:
-     *              0 : Needs to return a string
-     *              1 : Returns null string with unchecked task
-     *              2 : Returns null string with checked task
-     * @return String which highlights what Duke processed
-     */
-    public static String runTodo(ArrayList<Task> data, String input, TaskList.TaskState state) {
-        input = input.substring(5);
-        Task tempTask = new ToDo(input);
-        return getString(data, state, tempTask);
-    }
-
-    /**
      * Creates a new 'Deadline' task, before adding it to current list,
      * then returning the output by Duke.
      * @param data ArrayList of Tasks that's currently being stored
@@ -149,56 +127,6 @@ public class TaskList {
         }
         Task tempTask = new Deadline(taskInput, argumentDate);
         return getString(data, state, tempTask);
-    }
-
-    /**
-     * Method to run recurring tasks.
-     * @param data ArrayList of Tasks that's currently being stored
-     * @param input Command input by user
-     * @param state The type of output needed:
-     *              0 : Needs to return a string
-     *              1 : Returns null string with unchecked task
-     *              2 : Returns null string with checked task
-     * @param freq daily, weekly or monthly
-     * @return String which highlights what Duke processed
-     * @throws CakeException Shows error when cannot parse date
-     */
-    public static String runRecurring(ArrayList<Task> data, String input,
-                                      TaskList.TaskState state, String freq) throws CakeException {
-        input = input.substring(5).trim();
-        String tt1;
-        String tt2;
-        int token;
-        token = input.indexOf("/");
-        tt1 = input.substring(0, token - 1);
-        if (freq.equals("daily")) {
-            tt2 = input.substring(token + 7);
-        } else if (freq.equals("weekly")) {
-            tt2 = input.substring(token + 8);
-        } else {
-            tt2 = input.substring(token + 9);
-        }
-
-        // parse date here
-        Date startDate = parseDate(tt2);
-        Task tempTask = new RecurringTask(tt1, startDate, freq);
-        return getString(data, state, tempTask);
-    }
-
-    /**
-     * Method to return parsed Date.
-     * @param tt2 String to be parsed into Date
-     * @return Date parsed from the string
-     * @throws CakeException If date cannot be parsed
-     */
-    public static Date parseDate(String tt2) throws CakeException {
-        try {
-            com.joestelmach.natty.Parser parser = new com.joestelmach.natty.Parser();
-            List<DateGroup> groups = parser.parse(tt2);
-            return groups.get(0).getDates().get(0);
-        } catch (Exception e) {
-            throw new CakeException("[!] Date cannot be parsed: " + tt2);
-        }
     }
 
     private static String getString(ArrayList<Task> data, TaskList.TaskState state, Task tempTask) {
