+++ src/main/java/javacake/tasks/Deadline.java
@@ -1,14 +1,11 @@
 package javacake.tasks;
 
-import com.joestelmach.natty.DateGroup;
-import com.joestelmach.natty.Parser;
 import javacake.exceptions.CakeException;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 public class Deadline extends Task {
