+++ src/main/java/javacake/ui/AvatarScreen.java
@@ -3,15 +3,10 @@ package javacake.ui;
 import javafx.animation.Animation;
 import javafx.animation.KeyFrame;
 import javafx.animation.Timeline;
-import javafx.beans.Observable;
-import javafx.beans.value.ObservableValue;
-import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
-import javafx.scene.control.DialogPane;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
-import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.VBox;
 import javafx.util.Duration;
@@ -25,8 +20,6 @@ public class AvatarScreen extends VBox {
     private GridPane gridAvatar;
     @FXML
     private ImageView avatarImage;
-
-    private boolean isAvatarSwitched = false;
 
     public enum AvatarMode {
         HAPPY, EXTHAPPY, SAD, POUT
@@ -49,7 +42,6 @@ public class AvatarScreen extends VBox {
     private Image avatarPout2 = new Image(this.getClass().getResourceAsStream(
             "/images/avatar/poutclose.png"));
     List<Image> images = new ArrayList<>();
-    private int imageIndex = 0;
     private int timeFrame = 0;
     public static AvatarMode avatarMode;
 
@@ -78,7 +70,6 @@ public class AvatarScreen extends VBox {
             e.printStackTrace();
         }
         avatarImage.fitHeightProperty().bind(gridAvatar.heightProperty());
-        //avatarImage.setPreserveRatio(true);
 
         initialiseList();
         avatarMode = type;
