+++ src/main/java/javacake/ui/Ui.java
@@ -3,11 +3,9 @@ package javacake.ui;
 import javacake.Logic;
 import javacake.commands.ListNoteCommand;
 import javacake.commands.ScoreCommand;
-import javacake.quiz.QuizSession;
 import javacake.commands.ReminderCommand;
 import javacake.exceptions.CakeException;
 import javacake.quiz.Question;
-import javacake.storage.Profile;
 import javacake.storage.StorageManager;
 
 import java.io.BufferedReader;
@@ -121,23 +119,6 @@ public class Ui {
      */
     public void showMessage(String m) {
         System.out.print(m);
-    }
-
-    /**
-     * Method to display text from file.
-     * @param reader BufferedReader to read in text from file
-     * @throws CakeException Error thrown when unable to close reader
-     */
-    public void displayTextFile(BufferedReader reader) throws CakeException {
-        String lineBuffer;
-        try {
-            while ((lineBuffer = reader.readLine()) != null) {
-                System.out.println(lineBuffer);
-            }
-            reader.close();
-        } catch (IOException e) {
-            throw new CakeException("File not found!");
-        }
     }
 
     /**
