+++ src/main/java/javacake/unused/RecurringTask.java
@@ -1,4 +1,4 @@
-package javacake.tasks;
+package javacake.unused;
 
 import javacake.exceptions.CakeException;
